Change grapheme count to char count for highlighting
Renamed `num_graphemes_from_bytes` to `num_chars_from_bytes`

`char_map` of Token now holds the number of bytes in normalized
(deunicoded) string for each character of original string.

When `char_map` hasn't been initialized and we still try to call
`num_chars_from_bytes`, it uses the word that's currently in the Token
to count number of characters. Hence, it will be number of characters in
current string for number of bytes in current string.
Added tests for this case too.