Consider a char even if the bytes cover only some of it
Made the implementation of num_chars_from_bytes much simpler:
- Iterate over the char_indices until it finds index greater than num_bytes

For cases where a char is more than one byte, but we don't have enough bytes
left to complete the char, the char is considered regardless. This will
be useful for highlighting as when the search query is "gobriefc",
highlighting "Go💼" would be appropriate.

Also mentioned this in the doc string for the function and modified the
tests to confirm this behaviour.