Split camelCase in Latin segmenter
To improve recall and to be consistent with snake_case and kebab-case
splitting that's already in place, make the Latin Segmenter split words
on camelCase boundaries.

Define camelCase boundary as a lowercase letter directly followed by an
uppercase one. (Or the position between them, to be precise.) This
treats most cases and avoids common pitfalls like eg. ALL_CAPS.

Leverage the Unicode General Categories
https://en.wikipedia.org/wiki/Unicode_character_property#General_Category
and their support in the Regex crate for matching lowercase and
uppercase letters.

Put the logic into a separate module and expose API similar to
UnicodeSegmentation's to keep the call-site in latin.rs clean and
concise.

Caveats:
- Especially in code, it's common to write abbreviations in all caps.
For instance "meiliAPIClient". With this implementation it's split into
["meili", "APIClient"].
- The implementation is not grapheme-cluster aware. E.g. it won't split
on boundaries that are intermitted by combining characters like u{0306}.

Closes #129.