Compute boundary position manually.
The approach of using named empty capture group to pinpoint exact
location of the boundary required using the Regex's `captures` method
which allocated on the heap worsening performance.

Remove the capture group and compute the boundary position on our side.

Add a test for non-ASCII letters in the the boundary.

This further lowered the performance regression from ~14 % to ~5 %.