Improve Arabic language segmentation
Arabic text is segmented by word boundaries and by punctuation.
We need a workaround to segment the Arabic text that starts with `ال` (the) because it is not segmented by word boundaries.
One possible solution is to segment any word that starts with `ال` into two words. The `ال` and the rest of the word.
with this solution, we will have `الشجرة` (the tree) segmented into `ال` (the) and `شجرة` (tree). and if we search for `شجرة` (tree) or `الشجرة` (thetree) we will find results.
Some Arabic text starts with `ال` and not meant to be (the) like `البانيا` (Albania). In this case, we will have `ال` and `بانيا` segmented. and if we search for `البانيا` we will find results.