Fix pre-segmenter when a string start by an uncategorized character
When a field start by an uncategorized character like numbers these character will be put in a separated group than the first categorized character which gives a different segmentation on a same word.
For example a field containing `Memory 96GB` will be split as `["Memory", " ", "96GB"]` and a query containing `96GB` will be split as `["96", "GB"]` because `96` is uncategorized where `GB` is `Latin`. No we ensure to group every uncategorized character with the next script.